requirements.txt
pandas>=1.3.0
PyYAML 
flashtext  